Refactor MLA LoRA correction around routed matmul
Extract the explicit token-routing math into a reusable two-stage LoRA primitive and keep MLA-specific code limited to logical head views.
