+++ vllm/lora/ops/triton_ops/__init__.py
@@ -22,6 +22,7 @@ from vllm.lora.ops.triton_ops.mla_kv_b_lora import (
     mla_kv_b_lora_q,
     mla_kv_b_lora_v,
 )
+from vllm.lora.ops.triton_ops.routed_lora_matmul import routed_lora_two_stage
 
 __all__ = [
     "lora_expand",
@@ -32,6 +33,7 @@ __all__ = [
     "mla_kv_b_lora_linear",
     "mla_kv_b_lora_q",
     "mla_kv_b_lora_v",
+    "routed_lora_two_stage",
     "fused_moe_lora",
     "fused_moe_lora_shrink",
     "fused_moe_lora_expand",
